@@ -1,15 +1,17 @@
-apply plugin: 'com.android.application'
+plugins {
+    id 'com.android.application'
+    id 'kotlin-android'
+}
 
 android {
-    compileSdkVersion 23
-    buildToolsVersion "23.0.1"
+    compileSdkVersion 31
 
     defaultConfig {
         applicationId "com.miguelcatalan.materialsearchview.sample"
-        minSdkVersion 14
-        targetSdkVersion 23
+        minSdkVersion 22
+        targetSdkVersion 31
         versionCode 6
-        versionName "1.3.1"
+        versionName "1.5.0"
     }
     buildTypes {
         release {
@@ -20,8 +22,8 @@ android {
 }
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile project(':library')
-    compile 'com.android.support:appcompat-v7:23.1.0'
-    compile 'com.android.support:design:23.1.0'
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    implementation project(':library')
+    implementation "androidx.appcompat:appcompat:1.4.0"
+    implementation 'com.google.android.material:material:1.4.0'
 }
