@@ -1,10 +1,11 @@
 package com.miguelcatalan.materialsearchview.sample;
 
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import androidx.fragment.app.Fragment;
 
 /**
  * @author Miguel Catalan Bañuls
