@@ -2,12 +2,12 @@ package com.miguelcatalan.materialsearchview.sample;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.Button;
 
-public class MainActivity extends AppCompatActivity implements View.OnClickListener {
+import androidx.appcompat.app.AppCompatActivity;
 
+public class MainActivity extends AppCompatActivity implements View.OnClickListener {
     private Button defaultButton;
     private Button themedButton;
     private Button voiceButton;
